requirements.txt
PyYAML==6.0.2
Requests==2.31.0
termtables==0.2.4
